requirements.txt
asgiref==3.2.3
Django==3.0.2
numpy==1.18.1
pandas==1.0.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
xgboost==0.90
